feat: Реализован новый класс Smart. - В него перенесена логика updateElement & updateData. - Создан метод класса restoreHandlers, который добовляет трипу потерянные обработчики.
